Move mask logic to separate module
And cross-reference similar files so we don't miss it next time\!
